+++ tox.ini
@@ -6,7 +6,7 @@ skip_missing_interpreters = True
 
 commands =
     python --version
-    py.test --pyargs ihatemoney.tests.tests
+    py.test --pyargs ihatemoney.tests
 
 deps =
     -e.[dev]
